bugfix: check length of fields output on autocomplete
It seems there are some invisible non-whitespace characters which can end up at
the start of a line. This happened when mouse-copying text from a google doc.
Autocomplete does not recognize this as whitespace and so tries to complete it.
But string.Fields sees it as whitespace and returns no fields. This causes and
index-out-of-range error. Solution is to check fields before requesting
autocompletion.

	modified:   file/edit.go